pyproject.toml
[project]
name = "jenkins"
version = "0.1.0"
description = "A simple MCP server"
authors = [
    { name = "Michael Pan", email = "xhuaustc@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0.2",
    "typing-extensions>=4.14.0",
    "pytest-cov>=6.2.1",
    "requests>=2.32.4",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
backend-path = ["."]

[project.scripts]
"jenkins" = "jenkins.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "isort>=5.0", "mypy>=1.0", "ruff>=0.11.10"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
